minor changes to reduce duplications
