+++ backend/api/deployment_helper.py
@@ -1,26 +1,17 @@
 import logging
 import uuid
-from functools import wraps
 from typing import Any, Optional
 from urllib.parse import urlencode
 
+from api.api_key_validator import BaseAPIKeyValidator
 from api.constants import ApiExecution
-from api.exceptions import (
-    ApiKeyCreateException,
-    APINotFound,
-    Forbidden,
-    InactiveAPI,
-    UnauthorizedKey,
-)
+from api.exceptions import ApiKeyCreateException, APINotFound, Forbidden, InactiveAPI
 from api.key_helper import KeyHelper
 from api.models import APIDeployment, APIKey
 from api.serializers import APIExecutionResponseSerializer
 from django.core.files.uploadedfile import UploadedFile
 from django.db import connection
-from django_tenants.utils import get_tenant_model, tenant_context
-from rest_framework import status
 from rest_framework.request import Request
-from rest_framework.response import Response
 from rest_framework.serializers import Serializer
 from rest_framework.utils.serializer_helpers import ReturnDict
 from workflow_manager.endpoint.destination import DestinationConnector
@@ -32,61 +23,24 @@ from workflow_manager.workflow.workflow_helper import WorkflowHelper
 logger = logging.getLogger(__name__)
 
 
-class DeploymentHelper:
+class DeploymentHelper(BaseAPIKeyValidator):
     @staticmethod
-    def validate_api_key(func: Any) -> Any:
-        """Decorator that validates the API key.
+    def validate_specific_parameters(request: Request, **kwargs: Any) -> None:
+        """Validate api_name for API deployments."""
+        api_name = kwargs.get("api_name") or request.data.get("api_name")
+        if not api_name:
+            raise Forbidden("Missing api_name")
 
-        Sample header:
-            Authorization: Bearer 123e4567-e89b-12d3-a456-426614174001
-        Args:
-            func (Any): Function to wrap for validation
-        """
-
-        @wraps(func)
-        def wrapper(self: Any, request: Request, *args: Any, **kwargs: Any) -> Any:
-            """Wrapper to validate the inputs and key.
-
-            Args:
-                request (Request): Request context
-
-            Raises:
-                Forbidden: _description_
-                APINotFound: _description_
-
-            Returns:
-                Any: _description_
-            """
-            try:
-                authorization_header = request.headers.get("Authorization")
-                api_key = None
-                if authorization_header and authorization_header.startswith("Bearer "):
-                    api_key = authorization_header.split(" ")[1]
-                if not api_key:
-                    raise Forbidden("Missing api key")
-                org_name = kwargs.get("org_name") or request.data.get("org_name")
-                api_name = kwargs.get("api_name") or request.data.get("api_name")
-                if not api_name:
-                    raise Forbidden("Missing api_name")
-                tenant = get_tenant_model().objects.get(schema_name=org_name)
-                with tenant_context(tenant):
-                    api_deployment = DeploymentHelper.get_deployment_by_api_name(
-                        api_name=api_name
-                    )
-                    DeploymentHelper.validate_api(
-                        api_deployment=api_deployment, api_key=api_key
-                    )
-                    kwargs["api"] = api_deployment
-                    return func(self, request, *args, **kwargs)
-            except (UnauthorizedKey, InactiveAPI, APINotFound):
-                raise
-            except Exception as exception:
-                logger.error(f"Exception: {exception}")
-                return Response(
-                    {"error": str(exception)}, status=status.HTTP_403_FORBIDDEN
-                )
-
-        return wrapper
+    @staticmethod
+    def validate_and_process(
+        self: Any, request: Request, func: Any, api_key: str, *args: Any, **kwargs: Any
+    ) -> Any:
+        """Fetch API deployment and validate API key."""
+        api_name = kwargs.get("api_name") or request.data.get("api_name")
+        api_deployment = DeploymentHelper.get_deployment_by_api_name(api_name=api_name)
+        DeploymentHelper.validate_api(api_deployment=api_deployment, api_key=api_key)
+        kwargs["api"] = api_deployment
+        return func(self, request, *args, **kwargs)
 
     @staticmethod
     def validate_api(api_deployment: Optional[APIDeployment], api_key: str) -> None:
@@ -104,7 +58,7 @@ class DeploymentHelper:
             raise APINotFound()
         if not api_deployment.is_active:
             raise InactiveAPI()
-        KeyHelper.validate_api_key(api_key=api_key, api_instance=api_deployment)
+        KeyHelper.validate_api_key(api_key=api_key, instance=api_deployment)
 
     @staticmethod
     def validate_and_get_workflow(workflow_id: str) -> Workflow:
